# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem Mahad (Absensi, Hafalan, dll)",
        "title": "Mahad Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/absensi": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint ini digunakan untuk membuat absensi baru untuk Mahasantri berdasarkan data yang dikirimkan oleh mentor.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Absensi"
                ],
                "summary": "Membuat absensi baru untuk Mahasantri",
                "parameters": [
                    {
                        "description": "Data Absensi",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AbsensiRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Absensi created successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or Absensi already recorded for this time",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Mahasantri not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to create absensi",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/absensi/mahasantri/{mahasantri_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint ini digunakan untuk mendapatkan daftar absensi dari Mahasantri tertentu berdasarkan filter tanggal, waktu, status, serta mendukung paginasi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Absensi"
                ],
                "summary": "Mendapatkan daftar absensi berdasarkan Mahasantri ID dengan filter tanggal, waktu, dan status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mahasantri ID",
                        "name": "mahasantri_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tanggal awal filter (format: dd-mm-yyyy)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tanggal akhir filter (format: dd-mm-yyyy)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Waktu for filtering absensi",
                        "name": "waktu",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status for filtering absensi",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit number of results per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Absensi retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input or query parameters",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Mahasantri not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve absensi",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/absensi/mahasantri/{mahasantri_id}/daily-summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil data absensi harian Mahasantri selama 1 bulan berdasarkan waktu shubuh dan isya. Data akan mengisi status absen per hari, default \"belum-absen\" jika belum mengisi.",
                "tags": [
                    "Absensi"
                ],
                "summary": "Mendapatkan ringkasan absensi harian Mahasantri",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Mahasantri",
                        "name": "mahasantri_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bulan (format: MM, contoh: 04 untuk April)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tahun (format: YYYY, contoh: 2025)",
                        "name": "year",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil ringkasan absensi harian",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponseSwagger"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.AbsensiDailySummaryDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request - Format salah atau parameter tidak lengkap",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponseSwagger"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Gagal mengambil data absensi",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponseSwagger"
                        }
                    }
                }
            }
        },
        "/api/v1/absensi/mahasantri/{mahasantri_id}/per-month": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil total absensi bulanan berdasarkan waktu (shubuh \u0026 isya) dan status (hadir, izin, alpa) dalam satu bulan tertentu.",
                "tags": [
                    "Absensi"
                ],
                "summary": "Mendapatkan ringkasan absensi bulanan Mahasantri",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Mahasantri",
                        "name": "mahasantri_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bulan (format: MM, contoh: 04 untuk April)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tahun (format: YYYY, contoh: 2025)",
                        "name": "year",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil ringkasan absensi bulanan",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponseSwagger"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.AbsensiMonthlySummaryDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request - Format salah atau parameter tidak lengkap",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponseSwagger"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Gagal mengambil data absensi",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponseSwagger"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AbsensiDailySummaryDTO": {
            "type": "object",
            "properties": {
                "isya": {
                    "description": "hadir / absen / izin / belum-absen",
                    "type": "string"
                },
                "subuh": {
                    "description": "hadir / absen / izin / belum-absen",
                    "type": "string"
                },
                "tanggal": {
                    "description": "Format: dd-mm-yyyy",
                    "type": "string"
                }
            }
        },
...
        "dto.AbsensiRequestDTO": {
            "type": "object",
            "required": [
                "mahasantri_id",
                "status",
                "waktu"
            ],
            "properties": {
                "mahasantri_id": {
                    "description": "Validasi untuk memastikan MahasantriID ada",
                    "type": "integer"
                },
                "status": {
                    "description": "\"Hadir\", \"Absen\", atau \"Izin\"",
                    "type": "string",
                    "enum": [
                        "hadir",
                        "absen",
                        "izin"
                    ]
                },
                "waktu": {
                    "description": "\"Shubuh\" atau \"Isya\"",
                    "type": "string",
                    "enum": [
                        "shubuh",
                        "isya"
                    ]
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.SuccessResponseSwagger": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Request successful"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}